lane-dp: isolate lane logic in TTLaneInputBatch; delete lane_execution.py
Reduce lane-DP complexity per PR #399 review, keeping the generic execute
path free of mid-flow lane branches.

- De-dup: shared build_cached_request_state / apply_cached_req_state_update
  (input_batch.py), used by both the front-packed and lane state updates.
- TTLaneInputBatch now owns all lane-specific shaping + state:
  apply_step_plan (no-evict/no-condense state update), slot_block_tables /
  slot_sampling_params / slot_grammar_bitmask, build_model_input, and
  extract_output (merged-slot device/host read-back).
- execute_model dispatches once on the scheduler-attached TTStepPlan; removed
  execute_model_lanes and the worker's uses_tt_lane_coordinator branch (the
  worker is now lane-agnostic). Lane step orchestration lives in
  TTModelRunner._execute_lane_step (needs async_decode/submit_prefill/model).
- Collapsed the async lane wrappers: removed AsyncTTLaneModelRunnerOutput and
  complete_lane_decode_step; unified into AsyncTTModelRunnerOutput (optional
  scheduled_rows) + complete_decode_step.
- Deleted lane_execution.py / TTLaneStepExecutor.
- Tests migrated to the new seams; added merged-sampling isolation test
  (scheduled seeded row unaffected by an unscheduled random row; its RNG is
  not advanced).

Addresses review comments #3440428312, #3440244748, #3440510761, #3393716777.
