+++ plugins/vllm-tt-plugin/src/vllm_tt_plugin/input_batch.py
@@ -1,11 +1,13 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
-from typing import cast
+from typing import TYPE_CHECKING, Any, cast
 
 import numpy as np
 import torch
 
+from vllm.sampling_params import SamplingType
+from vllm.v1.outputs import LogprobsLists, LogprobsTensors
 from vllm.v1.sample.logits_processor import (
     BatchUpdateBuilder,
     LogitsProcessors,
@@ -19,6 +21,14 @@ from vllm.v1.sample.logits_processor.builtin import (
 from vllm.v1.sample.metadata import SamplingMetadata
 from vllm.v1.worker.block_table import MultiGroupBlockTable
 from vllm.v1.worker.gpu_input_batch import CachedRequestState
+from vllm_tt_plugin.logprobs import build_logprobs_from_topk
+from vllm_tt_plugin.model_input import TTModelInput, TTSamplingParams
+from vllm_tt_plugin.structured_output import reorder_grammar_bitmask_for_tt_batch
+
+if TYPE_CHECKING:
+    from vllm.v1.core.sched.output import GrammarOutput, SchedulerOutput
+    from vllm_tt_plugin.lane_scheduler import TTStepPlan
+    from vllm_tt_plugin.model_runner import TTModelRunner
 
 # Sentinel value for None seed. vLLM treats -1 as equivalent to None
 # (see SamplingParams.__post_init__), so we use -1 as the sentinel.
@@ -29,6 +39,62 @@ SEED_NONE_SENTINEL = -1
 # Can't use -1 because SamplingParams.logprobs=-1 means "all vocab logprobs".
 # although -1 gets remapped before writing to SamplingInputBatch.num_logprobs
 LOGPROBS_NONE_SENTINEL = -2
+
+
+def build_cached_request_state(new_req_data) -> CachedRequestState:
+    """Build a ``CachedRequestState`` for one newly-scheduled request.
+
+    Shared by the front-packed (``TTModelRunner._update_states``) and lane-DP
+    (``TTLaneInputBatch.apply_step_plan``) state updates so request-cache
+    construction -- prompt validation and per-request generator seeding -- lives
+    in exactly one place.
+    """
+    assert new_req_data.sampling_params is not None, (
+        "Pooling is not supported for TT yet"
+    )
+    if new_req_data.prompt_token_ids is None:
+        raise NotImplementedError("TT backend does not support prompt_embeds yet")
+    sampling_params = new_req_data.sampling_params
+    if sampling_params.sampling_type == SamplingType.RANDOM_SEED:
+        generator = torch.Generator(device="cpu")
+        generator.manual_seed(sampling_params.seed)
+    else:
+        generator = None
+    return CachedRequestState(
+        req_id=new_req_data.req_id,
+        prompt_token_ids=new_req_data.prompt_token_ids,
+        mm_features=new_req_data.mm_features,
+        sampling_params=sampling_params,
+        pooling_params=None,
+        generator=generator,
+        block_ids=new_req_data.block_ids,
+        num_computed_tokens=new_req_data.num_computed_tokens,
+        output_token_ids=[],
+        lora_request=new_req_data.lora_request,
+        prompt_embeds=new_req_data.prompt_embeds,
+    )
+
+
+def apply_cached_req_state_update(
+    req_state: CachedRequestState,
+    num_computed_tokens: int,
+    new_block_ids,
+    resumed_from_preemption: bool,
+) -> None:
+    """Apply a ``scheduled_cached_reqs`` update to a request's cached state.
+
+    Identical for front-packed and lane-DP: a request resumed from preemption
+    had its KV freed and rebuilt (replace block IDs), otherwise the newly
+    allocated blocks are appended. Persistent-batch row bookkeeping is left to
+    the caller.
+    """
+    req_state.num_computed_tokens = num_computed_tokens
+    if resumed_from_preemption:
+        assert new_block_ids is not None
+        req_state.block_ids = new_block_ids
+    elif new_block_ids is not None:
+        for block_ids, new_ids in zip(req_state.block_ids, new_block_ids):
+            block_ids.extend(new_ids)
 
 
 class SamplingInputBatch:
@@ -730,6 +796,86 @@ class TTLaneInputBatch(InputBatch):
         return
 
     # ------------------------------------------------------------------
+    # State update (lane step plan -> batch + request map)
+    # ------------------------------------------------------------------
+
+    def apply_step_plan(
+        self,
+        scheduler_output: "SchedulerOutput",
+        plan: "TTStepPlan",
+        requests: dict[str, CachedRequestState],
+        encoder_cache: dict,
+    ) -> bool:
+        """Apply one lane step plan to this batch and the runner's request map.
+
+        This is the lane-DP variant of ``TTModelRunner._update_states``. Unlike
+        the front-packed batch it does **not** evict merely-unscheduled
+        requests: a prefill step can leave running decodes unscheduled, and
+        freeing their stable device slot would disturb the on-device per-slot
+        seed RNG. Only finished requests, and requests resumed from preemption
+        (whose KV was rebuilt), release their slot. There is no condense.
+
+        ``requests`` (the runner's canonical ``req_id -> CachedRequestState``
+        map) and ``encoder_cache`` are mutated in place. Placement rows come
+        from ``plan.req_id_to_row`` -- the scheduler owns slot allocation, this
+        batch only materializes it. Returns whether the decode layout changed
+        (a placement or a freed slot), which the caller uses to reset the device
+        decode batch.
+        """
+        layout_changed = False
+
+        # Finished requests release their slot.
+        for req_id in scheduler_output.finished_req_ids:
+            requests.pop(req_id, None)
+            if self.remove_request(req_id) is not None:
+                layout_changed = True
+
+        # Free cached encoder outputs.
+        for mm_hash in scheduler_output.free_encoder_mm_hashes:
+            encoder_cache.pop(mm_hash, None)
+
+        req_ids_to_add: list[str] = []
+        for new_req_data in scheduler_output.scheduled_new_reqs:
+            req_id = new_req_data.req_id
+            requests[req_id] = build_cached_request_state(new_req_data)
+            req_ids_to_add.append(req_id)
+
+        # Running / resumed requests.
+        req_data = scheduler_output.scheduled_cached_reqs
+        for i, req_id in enumerate(req_data.req_ids):
+            req_state = requests[req_id]
+            num_computed_tokens = req_data.num_computed_tokens[i]
+            new_block_ids = req_data.new_block_ids[i]
+            resumed_from_preemption = req_id in req_data.resumed_req_ids
+            apply_cached_req_state_update(
+                req_state, num_computed_tokens, new_block_ids, resumed_from_preemption
+            )
+            if resumed_from_preemption:
+                # KV was freed and is being rebuilt; re-add fresh (drop the
+                # stale slot first). The slot may differ afterwards --
+                # acceptable under the exceptional preemption path, which
+                # re-prefills the request anyway.
+                if self.remove_request(req_id) is not None:
+                    layout_changed = True
+                req_ids_to_add.append(req_id)
+                continue
+            req_index = self.req_id_to_index.get(req_id)
+            if req_index is None:
+                req_ids_to_add.append(req_id)
+                continue
+            self.num_computed_tokens_cpu[req_index] = num_computed_tokens
+            if new_block_ids is not None:
+                self.block_table.append_row(new_block_ids, req_index)
+
+        # Place new / resumed requests at scheduler-owned stable rows.
+        for req_id in req_ids_to_add:
+            self.add_request_to_row(requests[req_id], plan.req_id_to_row[req_id])
+            layout_changed = True
+
+        self.refresh_logitsprocs()
+        return layout_changed
+
+    # ------------------------------------------------------------------
     # Sampling layout (merged, over the full slot batch)
     # ------------------------------------------------------------------
 
@@ -849,3 +995,381 @@ class TTLaneInputBatch(InputBatch):
             bad_words_token_ids=dict(sampling.bad_words_token_ids),
             logitsprocs=sampling.logitsprocs,
         )
+
+    # ------------------------------------------------------------------
+    # Per-row tensor views (the lane batch owns its own slicing/padding)
+    # ------------------------------------------------------------------
+
+    def slot_block_tables(
+        self, rows: list[int], zero_gaps: bool, total: int, width: int
+    ) -> list[torch.Tensor]:
+        """Per-group block tables for ``rows`` (one row per slot), each padded
+        to ``width`` (``max_num_blocks_per_req``). When ``zero_gaps`` is set,
+        rows of ``range(total)`` not in ``rows`` are zeroed (empty decode slots
+        carry no blocks)."""
+        occupied = set(rows)
+        out: list[torch.Tensor] = []
+        for bt in self.block_table.block_tables:
+            sel = list(range(total)) if zero_gaps else rows
+            bt_cpu = bt.get_cpu_tensor()[sel, :width].clone()
+            if bt_cpu.shape[1] < width:
+                pad = torch.zeros(
+                    bt_cpu.shape[0], width - bt_cpu.shape[1], dtype=bt_cpu.dtype
+                )
+                bt_cpu = torch.cat([bt_cpu, pad], dim=1)
+            if zero_gaps and len(occupied) < total:
+                gap = torch.ones(total, dtype=torch.bool)
+                gap[list(occupied)] = False
+                bt_cpu[gap] = 0
+            out.append(bt_cpu.contiguous())
+        return out
+
+    def slot_sampling_params(self, rows: list[int]) -> TTSamplingParams:
+        """Slice the slot-ordered sampling tensors to ``rows``."""
+        sp = self.sampling
+        idx = torch.as_tensor(rows, dtype=torch.long)
+        num_logprobs = sp.num_logprobs[idx]
+        return TTSamplingParams(
+            temperature=sp.temperature[idx],
+            top_k=sp.top_k[idx],
+            top_p=sp.top_p[idx],
+            presence_penalty=sp.presence_penalty[idx],
+            frequency_penalty=sp.frequency_penalty[idx],
+            repetition_penalty=sp.repetition_penalty[idx],
+            seed=sp.seed[idx],
+            num_logprobs=num_logprobs,
+            enable_log_probs=num_logprobs >= 0,
+        )
+
+    def slot_grammar_bitmask(
+        self, grammar_output: "GrammarOutput | None", batch_length: int
+    ) -> torch.Tensor | None:
+        """Reorder the scheduler grammar bitmask into a full slot-batch tensor.
+
+        Each structured-output request's bitmask row is placed at its device
+        slot row; every other slot is left all-ones (all tokens allowed), so the
+        mask lines up with the full slot logits the host sampler reads.
+        """
+        if grammar_output is None or grammar_output.grammar_bitmask is None:
+            return None
+        bitmask = torch.from_numpy(grammar_output.grammar_bitmask)
+        return reorder_grammar_bitmask_for_tt_batch(
+            bitmask=bitmask,
+            structured_output_request_ids=grammar_output.structured_output_request_ids,
+            req_id_to_index=self.req_id_to_index,
+            req_indices=list(range(batch_length)),
+            batch_length=batch_length,
+        )
+
+    # ------------------------------------------------------------------
+    # Merged model-input assembly (decode/prefill) and output extraction
+    # ------------------------------------------------------------------
+    #
+    # The lane batch materializes the whole merged device input from its own
+    # slot layout and reads the merged device output back into per-request
+    # tokens. ``runner`` is passed in for the genuinely runner/model-owned
+    # helpers (KV-layer block tables, device-sampling policy, multi-modal
+    # gather, the host sampler); everything row/slot-shaped lives here.
+
+    def build_model_input(
+        self,
+        runner: "TTModelRunner",
+        scheduler_output: "SchedulerOutput",
+        grammar_output: "GrammarOutput | None",
+        plan: "TTStepPlan",
+    ) -> TTModelInput:
+        """Build the merged device input for one lane step (decode or prefill)."""
+        if plan.is_decode:
+            return self._build_decode_input(
+                runner, scheduler_output, grammar_output, plan
+            )
+        return self._build_prefill_input(runner, scheduler_output, grammar_output, plan)
+
+    def _has_structured_outputs(
+        self,
+        runner: "TTModelRunner",
+        scheduler_output: "SchedulerOutput",
+        bitmask: torch.Tensor | None,
+    ) -> bool:
+        if bitmask is not None or scheduler_output.pending_structured_output_tokens:
+            return True
+        return any(
+            (req := runner.requests.get(req_id)) is not None
+            and req.sampling_params is not None
+            and req.sampling_params.structured_outputs is not None
+            for req_id in scheduler_output.num_scheduled_tokens
+        )
+
+    def _build_decode_input(
+        self,
+        runner: "TTModelRunner",
+        scheduler_output: "SchedulerOutput",
+        grammar_output: "GrammarOutput | None",
+        plan: "TTStepPlan",
+    ) -> TTModelInput:
+        """Build the merged decode input straight from the slot-ordered batch.
+
+        Every slot is present (gaps padded), so this is the device decode batch
+        with no scatter: row == device slot.
+        """
+        lane_batch = self
+        total = plan.capacity
+        occupied = lane_batch.occupied_rows()
+
+        num_tokens = lane_batch.num_tokens
+        positions_np = num_tokens[:total].astype(np.int32) - 1  # gaps -> -1
+        input_positions = torch.from_numpy(positions_np)
+        tokens_np = np.zeros((total, 1), dtype=np.int32)
+        for row in occupied:
+            tokens_np[row, 0] = lane_batch.token_ids_cpu[row, num_tokens[row] - 1]
+        input_tokens = torch.from_numpy(tokens_np)
+
+        block_tables_per_group = lane_batch.slot_block_tables(
+            occupied, zero_gaps=True, total=total, width=runner.max_num_blocks_per_req
+        )
+        rows_all = list(range(total))
+        tt_sampling_params = lane_batch.slot_sampling_params(rows_all)
+
+        bitmask = lane_batch.slot_grammar_bitmask(grammar_output, total)
+        has_structured = self._has_structured_outputs(runner, scheduler_output, bitmask)
+        perform_device_sampling = runner.check_perform_device_sampling(
+            is_decode=True, has_structured_outputs=has_structured
+        )
+
+        # The prompt/output token tensors feed device-side penalties only. Host
+        # sampling rebuilds them itself in ``build_merged_sampling_metadata``, so
+        # building them here too would be dead work on the host path.
+        prompt_tokens = output_tokens = None
+        if perform_device_sampling and not lane_batch.no_penalties:
+            prompt_tokens = lane_batch.make_prompt_token_ids_tensor(rows_all)
+            output_tokens = lane_batch.make_output_token_ids_tensor(rows_all)
+        reset_batch = runner._decode_layout_changed_since_last_decode
+        runner._decode_layout_changed_since_last_decode = False
+        slot_remap = lane_batch.pop_slot_remap()  # identity for stable slots
+
+        return TTModelInput(
+            input_tokens=input_tokens,
+            input_positions=input_positions,
+            prompt_lens=None,
+            block_tables=block_tables_per_group[0],
+            block_tables_per_group=block_tables_per_group,
+            block_tables_per_layer=runner._block_tables_per_layer(
+                block_tables_per_group
+            ),
+            # Device decodes every slot; only used for the empty-batch guard.
+            unpadded_batch_size=list(plan.batch_size_per_dp),
+            tt_sampling_params=tt_sampling_params,
+            multi_modal_kwargs={},
+            perform_device_sampling=perform_device_sampling,
+            grammar_bitmask=[bitmask],
+            prompt_tokens=prompt_tokens,
+            output_tokens=output_tokens,
+            reset_batch=reset_batch,
+            slot_remap=slot_remap,
+            # Host sampling reads the merged batch directly (see
+            # ``extract_output``); the per-rank sidecars are unused here.
+            allowed_token_ids_mask_list=[None],
+            bad_words_token_ids_list=[{}],
+            max_num_logprobs=[lane_batch.max_num_logprobs],
+            logitsprocs_list=[None],
+            generators_list=[{}],
+            prefill_empty_slots=None,
+        )
+
+    def _build_prefill_input(
+        self,
+        runner: "TTModelRunner",
+        scheduler_output: "SchedulerOutput",
+        grammar_output: "GrammarOutput | None",
+        plan: "TTStepPlan",
+    ) -> TTModelInput:
+        """Build the prefill input for the requests scheduled this step.
+
+        Prefill rows are front-packed in scheduler plan order. The plan carries
+        the stable slots (``prefill_empty_slots``) so ``submit_prefill`` seeds
+        each user at the device row decode will later read from. The output is
+        one token per prefilled request, in this same order.
+        """
+        lane_batch = self
+        rows = list(plan.input_rows)
+        rows_np = np.asarray(rows, dtype=np.int64)
+        input_positions = torch.from_numpy(
+            lane_batch.num_computed_tokens_cpu[rows_np].astype(np.int32)
+        )
+        prompt_lens = lane_batch.num_tokens[rows_np]
+        max_prefill = int(prompt_lens.max())
+        input_tokens = lane_batch.token_ids_cpu_tensor[rows_np, :max_prefill]
+
+        block_tables_per_group = lane_batch.slot_block_tables(
+            rows, zero_gaps=False, total=0, width=runner.max_num_blocks_per_req
+        )
+        tt_sampling_params = lane_batch.slot_sampling_params(rows)
+        batch_size_per_dp = list(plan.batch_size_per_dp)
+
+        bitmask = lane_batch.slot_grammar_bitmask(
+            grammar_output, lane_batch.max_num_reqs
+        )
+        has_structured = self._has_structured_outputs(runner, scheduler_output, bitmask)
+        perform_device_sampling = runner.check_perform_device_sampling(
+            is_decode=False, has_structured_outputs=has_structured
+        )
+
+        # Device-side penalties only; host sampling rebuilds these in
+        # ``build_merged_sampling_metadata`` (over the full slot batch), so
+        # building them here on the host path would be dead work.
+        prompt_tokens = output_tokens = None
+        if perform_device_sampling and not lane_batch.no_penalties:
+            prompt_tokens = lane_batch.make_prompt_token_ids_tensor(rows)
+            output_tokens = lane_batch.make_output_token_ids_tensor(rows)
+
+        multi_modal_kwargs = (
+            runner._gather_multi_modal_inputs(req_indices=list(rows))
+            if runner.model_config.is_multimodal_model
+            else {}
+        )
+
+        return TTModelInput(
+            input_tokens=input_tokens,
+            input_positions=input_positions,
+            prompt_lens=prompt_lens,
+            block_tables=block_tables_per_group[0],
+            block_tables_per_group=block_tables_per_group,
+            block_tables_per_layer=runner._block_tables_per_layer(
+                block_tables_per_group
+            ),
+            unpadded_batch_size=batch_size_per_dp,
+            tt_sampling_params=tt_sampling_params,
+            multi_modal_kwargs=multi_modal_kwargs,
+            perform_device_sampling=perform_device_sampling,
+            grammar_bitmask=[bitmask],
+            prompt_tokens=prompt_tokens,
+            output_tokens=output_tokens,
+            reset_batch=False,
+            slot_remap=None,
+            allowed_token_ids_mask_list=[None],
+            bad_words_token_ids_list=[{}],
+            max_num_logprobs=[lane_batch.max_num_logprobs],
+            logitsprocs_list=[None],
+            generators_list=[{}],
+            prefill_empty_slots=(
+                list(plan.prefill_empty_slots)
+                if plan.prefill_empty_slots is not None
+                else None
+            ),
+        )
+
+    def extract_output(
+        self,
+        runner: "TTModelRunner",
+        tt_out: Any,
+        tt_log_probs: Any,
+        model_input: TTModelInput,
+        scheduled_rows: list[int],
+        is_decode: bool,
+    ) -> tuple[torch.Tensor, LogprobsLists | None]:
+        """Read back one merged lane step into per-request sampled tokens.
+
+        Returns ``(sampled_token_ids[n, 1], logprobs)`` for the ``n``
+        ``scheduled_rows`` in order. Device sampling reads the sampled tokens
+        directly from each slot; host sampling runs **one** sampler call over
+        the whole slot batch (so the builtin/custom logits processors stay
+        row-aligned, with no per-lane slicing) and then picks the scheduled
+        rows out of the result. Also called from ``TTAsyncDecodeController`` to
+        finalize an async lane-decode step.
+        """
+        n = len(scheduled_rows)
+        rows_t = torch.as_tensor(scheduled_rows, dtype=torch.long)
+        if model_input.perform_device_sampling:
+            tokens = tt_out.reshape(-1) if isinstance(tt_out, torch.Tensor) else tt_out
+            # Decode reads each scheduled slot; prefill returns one token per
+            # scheduled request, already in row order.
+            sampled = tokens[rows_t] if is_decode else tokens[:n]
+            sampled = sampled.reshape(n, 1).to(torch.int32)
+            logprobs = self._device_logprobs(
+                tt_log_probs, model_input, scheduled_rows, sampled, is_decode
+            )
+            return sampled, logprobs
+
+        # Host sampling over the full slot batch.
+        total = self.max_num_reqs
+        logits = self._host_logits(tt_out, scheduled_rows, is_decode, total)
+        bitmask = model_input.grammar_bitmask[0]
+        if bitmask is not None:
+            runner.apply_grammar_bitmask(logits, bitmask)
+        sampling_metadata = self.build_merged_sampling_metadata(scheduled_rows)
+        sampler_output = runner.host_sampler(
+            logits=logits, sampling_metadata=sampling_metadata
+        )
+        sampled = sampler_output.sampled_token_ids.reshape(-1)[rows_t].reshape(n, 1)
+        logprobs = self._host_logprobs(sampler_output.logprobs_tensors, scheduled_rows)
+        return sampled.to(torch.int32), logprobs
+
+    def _host_logits(
+        self, tt_out: Any, scheduled_rows: list[int], is_decode: bool, total: int
+    ) -> torch.Tensor:
+        """Full ``[total, vocab]`` slot logits for host sampling.
+
+        Decode logits already cover every slot. Prefill logits cover only the
+        scheduled requests (row order), so scatter them onto their slot rows;
+        the unscheduled / gap rows are sampled harmlessly and dropped.
+        """
+        logits = tt_out[:, -1, :] if tt_out.dim() == 3 else tt_out
+        if is_decode:
+            return logits
+        full = torch.zeros((total, logits.shape[-1]), dtype=logits.dtype)
+        full[torch.as_tensor(scheduled_rows, dtype=torch.long)] = logits[
+            : len(scheduled_rows)
+        ]
+        return full
+
+    def _host_logprobs(
+        self, logprobs_tensors: LogprobsTensors | None, scheduled_rows: list[int]
+    ) -> LogprobsLists | None:
+        if logprobs_tensors is None:
+            return None
+        rows_t = torch.as_tensor(scheduled_rows, dtype=torch.long)
+        return LogprobsTensors(
+            logprob_token_ids=logprobs_tensors.logprob_token_ids[rows_t],
+            logprobs=logprobs_tensors.logprobs[rows_t],
+            selected_token_ranks=logprobs_tensors.selected_token_ranks[rows_t],
+        ).tolists()
+
+    def _device_logprobs(
+        self,
+        tt_log_probs: Any,
+        model_input: TTModelInput,
+        scheduled_rows: list[int],
+        sampled: torch.Tensor,
+        is_decode: bool,
+    ) -> LogprobsLists | None:
+        """Build logprobs for device-sampled tokens, mirroring the gather-DP
+        device logprobs path but over the scheduled slot rows."""
+        n = len(scheduled_rows)
+        assert isinstance(model_input.tt_sampling_params.enable_log_probs, torch.Tensor)
+        enable = model_input.tt_sampling_params.enable_log_probs
+        sel = (
+            torch.as_tensor(scheduled_rows, dtype=torch.long)
+            if is_decode
+            else torch.arange(n, dtype=torch.long)
+        )
+        if not enable[sel].any():
+            return None
+        assert tt_log_probs is not None, "model should return logprobs when requested"
+        max_lp = model_input.max_num_logprobs[0] or 0
+        next_token_ids = sampled.reshape(n)
+        if isinstance(tt_log_probs, tuple):
+            top_k_logprobs, top_k_indices = tt_log_probs
+            logprobs_tensors = build_logprobs_from_topk(
+                top_k_logprobs=top_k_logprobs[sel],
+                top_k_indices=top_k_indices[sel],
+                sampled_token_ids=next_token_ids,
+                max_num_logprobs=max_lp,
+            )
+        else:
+            sampled_log_probs = tt_log_probs.reshape(-1)[sel].reshape(n)
+            logprobs_tensors = LogprobsTensors(
+                logprob_token_ids=next_token_ids.unsqueeze(-1).to(torch.int32),
+                logprobs=sampled_log_probs.unsqueeze(-1).to(torch.float32),
+                selected_token_ranks=torch.full((n,), -1, dtype=torch.int32),
+            )
+        return logprobs_tensors.tolists()
