+++ plugins/vllm-tt-plugin/tests/test_lane_model_runner.py
@@ -1,10 +1,12 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-"""Host-only tests for the lane-DP model-runner adapter.
+"""Host-only tests for the lane-DP step seams.
 
-No device execution; the lane executor's index logic runs on plain tensors with
-fake runner/model collaborators. Requires the ttnn-enabled environment because
-importing the plugin modules pulls in ttnn.
+No device execution; the lane row/slot index logic runs on plain tensors with
+fake runner/model collaborators. Lane-specific input/output shaping lives on
+``TTLaneInputBatch`` (``extract_output``); the thin step orchestration lives on
+``TTModelRunner`` (``_execute_lane_step``). Requires the ttnn-enabled
+environment because importing the plugin modules pulls in ttnn.
 """
 
 from types import SimpleNamespace
@@ -17,10 +19,27 @@ from vllm_tt_plugin.async_decode import (
     TTDecodeSubmission,
     TTFinalizedDecode,
 )
-from vllm_tt_plugin.lane_execution import TTLaneStepExecutor
+from vllm_tt_plugin.input_batch import TTLaneInputBatch
 from vllm_tt_plugin.lane_scheduler import TTStepPlan
+from vllm_tt_plugin.model_runner import TTModelRunner
 
 from vllm.v1.core.sched.output import SchedulerOutput
+
+VOCAB = 8
+BLOCK = 16
+MAX_MODEL_LEN = 256
+
+
+def _lane_batch(num_lanes=1, per_lane=5):
+    return TTLaneInputBatch(
+        num_lanes=num_lanes,
+        per_lane=per_lane,
+        max_model_len=MAX_MODEL_LEN,
+        max_num_batched_tokens=MAX_MODEL_LEN * num_lanes * per_lane,
+        vocab_size=VOCAB,
+        block_sizes=[BLOCK],
+        kernel_block_sizes=[BLOCK],
+    )
 
 
 def _scheduler_output_with_plan(plan: TTStepPlan) -> SchedulerOutput:
@@ -43,9 +62,13 @@ def _capturing_host_sampler(captured: dict):
     return sampler
 
 
-def test_lane_executor_rejects_request_specific_rope():
+# --------------------------------------------------------------------------
+# Orchestration guard
+# --------------------------------------------------------------------------
+
+
+def test_lane_step_rejects_request_specific_rope():
     runner = SimpleNamespace(request_specific_rope=True)
-    executor = TTLaneStepExecutor(runner)
     output = _scheduler_output_with_plan(
         TTStepPlan(
             is_decode=True,
@@ -60,11 +83,16 @@ def test_lane_executor_rejects_request_specific_rope():
     )
 
     with pytest.raises(NotImplementedError, match="request-specific RoPE"):
-        executor.execute_model(output, grammar_output=None)
+        TTModelRunner._execute_lane_step(runner, output, grammar_output=None)
 
 
-def test_extract_step_device_decode_returns_scheduled_rows_in_order():
-    executor = TTLaneStepExecutor(SimpleNamespace())
+# --------------------------------------------------------------------------
+# TTLaneInputBatch.extract_output (merged device/host read-back)
+# --------------------------------------------------------------------------
+
+
+def test_extract_output_device_decode_returns_scheduled_rows_in_order():
+    batch = _lane_batch()
     # Full slot decode output: one sampled token per device slot.
     tt_out = torch.tensor([10, 11, 12, 13, 14], dtype=torch.int32)
     model_input = SimpleNamespace(
@@ -75,16 +103,21 @@ def test_extract_step_device_decode_returns_scheduled_rows_in_order():
         max_num_logprobs=[None],
     )
 
-    sampled, logprobs = executor.extract_step(
-        tt_out, None, model_input, scheduled_rows=[4, 1], is_decode=True
+    sampled, logprobs = batch.extract_output(
+        SimpleNamespace(),
+        tt_out,
+        None,
+        model_input,
+        scheduled_rows=[4, 1],
+        is_decode=True,
     )
 
     assert sampled.tolist() == [[14], [11]]  # slots 4 and 1, in scheduled order
     assert logprobs is None
 
 
-def test_extract_step_device_prefill_returns_front_packed_tokens():
-    executor = TTLaneStepExecutor(SimpleNamespace())
+def test_extract_output_device_prefill_returns_front_packed_tokens():
+    batch = _lane_batch()
     # Prefill returns one token per scheduled request, in plan order.
     tt_out = torch.tensor([20, 21], dtype=torch.int32)
     model_input = SimpleNamespace(
@@ -95,59 +128,61 @@ def test_extract_step_device_prefill_returns_front_packed_tokens():
         max_num_logprobs=[None],
     )
 
-    sampled, logprobs = executor.extract_step(
-        tt_out, None, model_input, scheduled_rows=[4, 1], is_decode=False
+    sampled, logprobs = batch.extract_output(
+        SimpleNamespace(),
+        tt_out,
+        None,
+        model_input,
+        scheduled_rows=[4, 1],
+        is_decode=False,
     )
 
     assert sampled.tolist() == [[20], [21]]
     assert logprobs is None
 
 
-def test_extract_step_host_decode_samples_full_slot_then_picks_scheduled_rows():
+def test_extract_output_host_decode_samples_full_slot_then_picks_rows():
     captured: dict = {}
-    runner = SimpleNamespace(
-        lane_batch=SimpleNamespace(
-            max_num_reqs=5, build_merged_sampling_metadata=lambda rows: rows
-        ),
-        host_sampler=_capturing_host_sampler(captured),
-    )
-    executor = TTLaneStepExecutor(runner)
-    # Full slot logits: row r's argmax is token r.
-    logits = torch.full((5, 8), -10.0)
+    batch = _lane_batch()
+    batch.build_merged_sampling_metadata = lambda rows: None  # sampler ignores it
+    runner = SimpleNamespace(host_sampler=_capturing_host_sampler(captured))
+    # Full slot logits: row r's argmax is token r (vocab>=5).
+    logits = torch.full((5, VOCAB), -10.0)
     for r in range(5):
         logits[r, r] = 5.0
     model_input = SimpleNamespace(perform_device_sampling=False, grammar_bitmask=[None])
 
-    sampled, logprobs = executor.extract_step(
-        logits, None, model_input, scheduled_rows=[4, 1], is_decode=True
+    sampled, logprobs = batch.extract_output(
+        runner, logits, None, model_input, scheduled_rows=[4, 1], is_decode=True
     )
 
-    assert captured["logits"].shape == (5, 8)  # sampled the whole slot batch once
+    assert captured["logits"].shape == (5, VOCAB)  # sampled the whole slot batch once
     assert sampled.tolist() == [[4], [1]]  # rows 4 and 1, in scheduled order
     assert logprobs is None
 
 
-def test_extract_step_host_prefill_scatters_logits_to_stable_rows():
+def test_extract_output_host_prefill_scatters_logits_to_stable_rows():
     captured: dict = {}
-    runner = SimpleNamespace(
-        lane_batch=SimpleNamespace(
-            max_num_reqs=5, build_merged_sampling_metadata=lambda rows: rows
-        ),
-        host_sampler=_capturing_host_sampler(captured),
-    )
-    executor = TTLaneStepExecutor(runner)
+    batch = _lane_batch()
+    batch.build_merged_sampling_metadata = lambda rows: None
+    runner = SimpleNamespace(host_sampler=_capturing_host_sampler(captured))
     # Prefill logits: one row per scheduled request (front-packed, plan order).
-    prefill_logits = torch.full((2, 8), -10.0)
+    prefill_logits = torch.full((2, VOCAB), -10.0)
     prefill_logits[0, 3] = 5.0  # scheduled request 0 -> token 3
     prefill_logits[1, 6] = 5.0  # scheduled request 1 -> token 6
     model_input = SimpleNamespace(perform_device_sampling=False, grammar_bitmask=[None])
 
-    sampled, _ = executor.extract_step(
-        prefill_logits, None, model_input, scheduled_rows=[4, 1], is_decode=False
+    sampled, _ = batch.extract_output(
+        runner,
+        prefill_logits,
+        None,
+        model_input,
+        scheduled_rows=[4, 1],
+        is_decode=False,
     )
 
     full = captured["logits"]
-    assert full.shape == (5, 8)  # scattered onto the full slot grid before sampling
+    assert full.shape == (5, VOCAB)  # scattered onto the full slot grid before sampling
     assert torch.equal(full[4], prefill_logits[0])  # request 0 -> stable row 4
     assert torch.equal(full[1], prefill_logits[1])  # request 1 -> stable row 1
     for gap in (0, 2, 3):
@@ -155,9 +190,12 @@ def test_extract_step_host_prefill_scatters_logits_to_stable_rows():
     assert sampled.tolist() == [[3], [6]]
 
 
-def test_submit_prefill_forwards_plan_empty_slots_to_model():
-    from vllm_tt_plugin.model_runner import TTModelRunner
+# --------------------------------------------------------------------------
+# Shared device submission / async lane decode
+# --------------------------------------------------------------------------
 
+
+def test_submit_prefill_forwards_plan_empty_slots_to_model():
     captured: dict = {}
 
     class FakeModel:
@@ -189,17 +227,19 @@ def test_submit_prefill_forwards_plan_empty_slots_to_model():
     assert captured["empty_slots"] == [4]
 
 
-def test_async_lane_decode_uses_lane_executor_extraction():
+def test_async_lane_decode_uses_batch_extraction():
     calls = []
 
-    class FakeLaneExecutor:
-        def extract_step(
-            self, tt_out, tt_log_probs, model_input, scheduled_rows, is_decode
+    class FakeLaneBatch:
+        def extract_output(
+            self, runner, tt_out, tt_log_probs, model_input, scheduled_rows, is_decode
         ):
-            calls.append((tt_out, tt_log_probs, model_input, scheduled_rows, is_decode))
+            calls.append(
+                (runner, tt_out, tt_log_probs, model_input, scheduled_rows, is_decode)
+            )
             return torch.tensor([[7]], dtype=torch.int32), None
 
-    runner = SimpleNamespace(lane_executor=FakeLaneExecutor())
+    runner = SimpleNamespace(lane_batch=FakeLaneBatch())
     controller = TTAsyncDecodeController(runner)
     controller.finalize_decode = lambda submission: TTFinalizedDecode(
         tt_out=torch.tensor([[1, 2, 3]], dtype=torch.float32),
@@ -220,7 +260,7 @@ def test_async_lane_decode_uses_lane_executor_extraction():
     )
     model_input = object()
 
-    completed = controller.complete_lane_decode_step(
+    completed = controller.complete_decode_step(
         submission=submission,
         model_input=model_input,
         context=context,
@@ -231,7 +271,10 @@ def test_async_lane_decode_uses_lane_executor_extraction():
     assert completed.logprobs is None
     assert completed.context is context
     assert len(calls) == 1
-    tt_out, tt_log_probs, seen_model_input, scheduled_rows, is_decode = calls[0]
+    seen_runner, tt_out, tt_log_probs, seen_model_input, scheduled_rows, is_decode = (
+        calls[0]
+    )
+    assert seen_runner is runner
     assert torch.equal(tt_out, torch.tensor([[1, 2, 3]], dtype=torch.float32))
     assert tt_log_probs is None
     assert seen_model_input is model_input
